Split test helpers into separate files
